+++ PaymentMethods/HeidelpayGiropayPaymentMethod.php
@@ -78,27 +78,15 @@ class HeidelpayGiropayPaymentMethod extends HeidelpayAbstractPaymentMethod
     }
 
     /**
-     * Returns the redirect url to the giropay site.
-     *
-     * @param $quote
-     * @return array $response An array with heidelpay processing results
+     * @inheritdoc
      */
     public function getHeidelpayUrl($quote)
     {
         parent::getHeidelpayUrl($quote);
 
-        // force PhpApi to just generate the request instead of sending it directly
-        $this->_heidelpayPaymentMethod->_dryRun = true;
-
         // set payment type to debit
         $this->_heidelpayPaymentMethod->authorize();
 
-        // pepare and send request to heidelpay
-        $response = $this->_heidelpayPaymentMethod->getRequest()->send(
-            $this->_heidelpayPaymentMethod->getPaymentUrl(),
-            $this->_heidelpayPaymentMethod->getRequest()->convertToArray()
-        );
-
-        return $response[0];
+        return $this->_heidelpayPaymentMethod->getResponse();
     }
 }